// src/features/posts/schemas/post.view.ts
class likes {
  addedAt: string;
  userId: string;
  login: string;
}

export class PostView {
  id: string;
  title: string;
  shortDescription: string;
  content: string;
  blogId: string;
  blogName: string;
  createdAt: string;
  extendedLikesInfo: {
    likesCount: number;
    dislikesCount: number;
    myStatus: string;
    newestLikes: likes[];
  };
}